Seperate the logics into middleware
